+++ src/main/java/com/boj/guidance/service/MemberService.java
@@ -1,9 +1,9 @@
 package com.boj.guidance.service;
 
-import com.boj.guidance.dto.MemberAuthRequestDto;
-import com.boj.guidance.dto.MemberJoinRequestDto;
-import com.boj.guidance.dto.MemberLoginRequestDto;
-import com.boj.guidance.dto.MemberResponseDto;
+import com.boj.guidance.dto.MemberDto.MemberAuthRequestDto;
+import com.boj.guidance.dto.MemberDto.MemberJoinRequestDto;
+import com.boj.guidance.dto.MemberDto.MemberLoginRequestDto;
+import com.boj.guidance.dto.MemberDto.MemberResponseDto;
 import org.springframework.transaction.annotation.Transactional;
 
 public interface MemberService {
